+++ src/config/GeneralConfig.php
@@ -143,11 +143,11 @@ class GeneralConfig extends BaseObject
     public bool $allowUpdates = true;
 
     /**
-     * @var string[]|string The file extensions Craft should allow when a user is uploading files.
+     * @var string[] The file extensions Craft should allow when a user is uploading files.
      * @see extraAllowedFileExtensions
      * @group Assets
      */
-    public string|array $allowedFileExtensions = [
+    public array $allowedFileExtensions = [
         '7z',
         'aiff',
         'asc',
@@ -340,6 +340,16 @@ class GeneralConfig extends BaseObject
     public ?string $brokenImagePath = null;
 
     /**
+     * @var string|null A unique ID representing the current build of the codebase.
+     *
+     * This should be set to something unique to the deployment, e.g. a Git SHA or a deployment timestamp.
+     *
+     * @since 4.0.0
+     * @group Environment
+     */
+    public ?string $buildId = null;
+
+    /**
      * @var mixed The default length of time Craft will store data, RSS feed, and template caches.
      *
      * If set to `0`, data and RSS feed caches will be stored indefinitely; template caches will be stored for one year.
@@ -405,7 +415,7 @@ class GeneralConfig extends BaseObject
      * the front-end website.
      *
      * This can be set to `null` if you have a dedicated host name for the control panel (e.g. `cms.example.com`), or you are running Craft in
-     * [Headless Mode](config3:headlessMode). If you do that, you will need to ensure that the control panel is being served from its own webroot
+     * [Headless Mode](config3:headlessMode). If you do that, you will need to ensure that the control panel is being served from its own web root
      * directory on your server, with an `index.php` file that defines the `CRAFT_CP` PHP constant.
      *
      * ```php
@@ -413,7 +423,7 @@ class GeneralConfig extends BaseObject
      * ```
      *
      * Alternatively, you can set the <config3:baseCpUrl> config setting, but then you will run the risk of losing access to portions of your
-     * control panel due to URI conflicts with actual folders/files in your main webroot.
+     * control panel due to URI conflicts with actual folders/files in your main web root.
      *
      * (For example, if you have an `assets/` folder, that would conflict with the `/assets` page in the control panel.)
      *
@@ -608,7 +618,7 @@ class GeneralConfig extends BaseObject
     public bool $enableCsrfCookie = true;
 
     /**
-     * @var bool Whether GraphQL introspection queries are allowed. Defaults to `true` and is always allowed in the CP.
+     * @var bool Whether GraphQL introspection queries are allowed. Defaults to `true` and is always allowed in the control panel.
      * @since 3.6.0
      * @group GraphQL
      */
@@ -682,11 +692,11 @@ class GeneralConfig extends BaseObject
     public string $errorTemplatePrefix = '';
 
     /**
-     * @var string[]|string|null List of file extensions that will be merged into the <config3:allowedFileExtensions> config setting.
+     * @var string[]|null List of file extensions that will be merged into the <config3:allowedFileExtensions> config setting.
      * @see allowedFileExtensions
      * @group System
      */
-    public string|array|null $extraAllowedFileExtensions = null;
+    public ?array $extraAllowedFileExtensions = null;
 
     /**
      * @var string[]|null List of extra locale IDs that the application should support, and users should be able to select as their Preferred Language.
@@ -966,7 +976,7 @@ class GeneralConfig extends BaseObject
     /**
      * @var bool Whether generated URLs should omit `index.php` (e.g. `http://domain.com/path` instead of `http://domain.com/index.php/path`)
      *
-     * This can only be possible if your server is configured to redirect would-be 404's to `index.php`, for example, with the redirect found
+     * This can only be possible if your server is configured to redirect would-be 404s to `index.php`, for example, with the redirect found
      * in the `.htaccess` file that came with Craft:
      *
      * ```
@@ -1591,7 +1601,7 @@ class GeneralConfig extends BaseObject
 
     /**
      * @var bool|string Determines what protocol/schema Craft will use when generating tokenized URLs. If set to `'auto'`, Craft will check the
-     * current site’s base URL and the protocol of the current request and if either of them are https will use `https` in the tokenized URL. If not,
+     * current site’s base URL and the protocol of the current request and if either of them are HTTPS will use `https` in the tokenized URL. If not,
      * will use `http`.
      *
      * If set to `false`, Craft will always use `http`. If set to `true`, then, Craft will always use `https`.
@@ -1708,12 +1718,6 @@ class GeneralConfig extends BaseObject
     public function init(): void
     {
         // Merge extraAllowedFileExtensions into allowedFileExtensions
-        if (is_string($this->allowedFileExtensions)) {
-            $this->allowedFileExtensions = StringHelper::split($this->allowedFileExtensions);
-        }
-        if (is_string($this->extraAllowedFileExtensions)) {
-            $this->extraAllowedFileExtensions = StringHelper::split($this->extraAllowedFileExtensions);
-        }
         if (is_array($this->extraAllowedFileExtensions)) {
             $this->allowedFileExtensions = array_merge($this->allowedFileExtensions, $this->extraAllowedFileExtensions);
             $this->extraAllowedFileExtensions = null;
@@ -1738,7 +1742,7 @@ class GeneralConfig extends BaseObject
         // Normalize size settings
         $this->maxUploadFileSize = ConfigHelper::sizeInBytes($this->maxUploadFileSize);
 
-        // Normalize the default CP language
+        // Normalize the default control panel language
         if (isset($this->defaultCpLanguage)) {
             try {
                 $this->defaultCpLanguage = Localization::normalizeLanguage($this->defaultCpLanguage);
@@ -1756,6 +1760,11 @@ class GeneralConfig extends BaseObject
                     throw new InvalidConfigException($e->getMessage(), 0, $e);
                 }
             }
+        }
+
+        // Normalize disabledPlugins
+        if (is_string($this->disabledPlugins) && $this->disabledPlugins !== '*') {
+            $this->disabledPlugins = StringHelper::split($this->disabledPlugins);
         }
     }