+++ src/services/Assets.php
@@ -600,7 +600,7 @@ class Assets extends Component
             }
         }
 
-        // If it's not an image, return a generic file extension icon
+        // If it’s not an image, return a generic file extension icon
         $extension = $asset->getExtension();
         if (!Image::canManipulateAsImage($extension) || !$asset->getVolume()->getTransformFs()->hasUrls) {
             return AssetsHelper::iconUrl($extension);